backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy>=1.24.0
Pillow>=10.0.0
ultralytics==8.1.9
easyocr==1.7.1
torch>=2.0.0
torchvision>=0.15.0
